+++ tests/testthat/test-bind.R
@@ -71,21 +71,27 @@ test_that("auto-renaming works", {
   )
 })
 
-# FIXME: this and "out/bind_src_mismatch_error.txt" can be removed, once the FIXME below is resolved
 test_that("test error output for src mismatches", {
   skip_if_not_installed("dbplyr")
-  verify_output("out/bind_src_mismatch_error.txt", dm_bind(dm_for_flatten(), dm_for_filter_sqlite()))
+
+  expect_snapshot({
+    writeLines(conditionMessage(expect_error(
+      dm_bind(dm_for_flatten(), dm_for_filter_sqlite())
+    )))
+  })
 })
 
-verify_output("out/bind.txt", {
-  dm_bind()
-  dm_bind(empty_dm())
-  dm_bind(dm_for_filter()) %>% collect()
-  dm_bind(dm_for_filter(), dm_for_flatten(), dm_for_filter(), repair = "unique") %>% collect()
-  dm_bind(dm_for_filter(), dm_for_flatten(), dm_for_filter(), repair = "unique", quiet = TRUE) %>% collect()
-  dm_bind(dm_for_filter(), dm_for_flatten(), dm_for_filter())
-  # FIXME: this would fail for TIC_ONLY_IMPORTS, cause {dbplyr} is missing; is there a way to skip in `verify_output()`?
-  # dm_bind(dm_for_flatten(), dm_for_filter_sqlite())
+test_that("output", {
+  expect_snapshot({
+    dm_bind()
+    dm_bind(empty_dm())
+    dm_bind(dm_for_filter()) %>% collect()
+    dm_bind(dm_for_filter(), dm_for_flatten(), dm_for_filter(), repair = "unique") %>% collect()
+    dm_bind(dm_for_filter(), dm_for_flatten(), dm_for_filter(), repair = "unique", quiet = TRUE) %>% collect()
+    writeLines(conditionMessage(expect_error(
+      dm_bind(dm_for_filter(), dm_for_flatten(), dm_for_filter())
+    )))
+  })
 })
 
 verify_output("out/compound-bind.txt", {